refactor(app): lift the CLI into extenddb-app; bin becomes the postgres thin bin
Completes the library decoupling. The full CLI (serve, init, destroy, verify,
migrate, status, stop, settings, manage, catalog-check) plus its dispatch,
subcommand modules, and helpers move out of crates/bin into a new,
backend-agnostic extenddb-app crate exposing `run(BuildInfo)`.

crates/bin is now the reference thin bin: it registers exactly one backend
(postgres), installs the registry, and calls extenddb_app::run. This is the
copy-paste template for a third-party backend author — swap the register call,
supply your own build provenance, ship an extenddb-<backend> image, and touch
no ExtendDB core crate. First-party and third-party backends now follow the
identical path; there is no privileged wiring for postgres.

Build provenance (git hash, build time) is passed in via BuildInfo rather than
read through env!(), since the app library cannot see the bin's build.rs
environment. cmd_serve::run takes git_hash and forwards it to
extenddb_server::serve. The bin's dependency set collapses to extenddb-app,
extenddb-storage, the backend crate, and anyhow.

Behavior-preserving: workspace unit tests green (incl. the serve arg-parsing
tests now in extenddb-app), clippy -D warnings + fmt clean.
