+++ crates/app/src/cmd_catalog_check.rs
