+++ crates/app/src/cmd_destroy.rs
