+++ crates/app/src/cmd_init.rs
