+++ crates/app/src/cmd_manage.rs
