+++ crates/app/src/cmd_migrate.rs
