+++ crates/app/src/cmd_settings.rs
