+++ crates/app/src/cmd_status.rs
