+++ crates/app/src/cmd_stop.rs
