+++ crates/app/src/cmd_verify.rs
