+++ crates/bin/Cargo.toml
@@ -16,27 +16,7 @@ default = ["postgres"]
 postgres = ["extenddb-storage-postgres"]
 
 [dependencies]
-extenddb-auth = { workspace = true }
-extenddb-cache = { workspace = true }
-extenddb-core = { workspace = true }
-extenddb-engine = { workspace = true }
+extenddb-app = { workspace = true }
 extenddb-storage = { workspace = true }
-extenddb-config = { workspace = true }
 extenddb-storage-postgres = { workspace = true, optional = true }
-extenddb-server = { workspace = true }
-tokio = { workspace = true }
 anyhow = { workspace = true }
-clap = { workspace = true }
-daemonize = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-sqlx = { workspace = true }
-time = { workspace = true }
-toml = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-base64 = { workspace = true }
-libc = { workspace = true }
-rcgen = { workspace = true }
-rustls = { workspace = true }
-rustls-pemfile = { workspace = true }
